requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
requests
python-telegram-bot
python-dotenv
pypdf
google-generativeai
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
torch